Added song appears in the list view

+++ app/src/main/java/com/example/android/musicapplicationv2/data/Song.kt
@@ -17,5 +17,5 @@ data class Song (
 
     var year: Int,
 
-    var duration: Long
+    var duration: Int
     )
